Move common FastAPI configuration code to webapps/base/api.py
